Decouple queue management and internal dependencies
