+++ packages/hasher/src/__tests__/resolveInternalDependencies.test.ts
@@ -1,13 +1,10 @@
-import * as path from "path";
 import { setupFixture } from "backfill-utils-test";
 
 import { filterDependenciesInFixture } from "./resolveDependenciesHelper";
 import { getYarnWorkspaces } from "../yarnWorkspaces";
-import { PackageHashInfo } from "../hashOfPackage";
 import {
   filterInternalDependencies,
-  resolveInternalDependencies,
-  addToQueue
+  resolveInternalDependencies
 } from "../resolveInternalDependencies";
 
 describe("filterInternalDependencies()", () => {
@@ -30,54 +27,6 @@ describe("filterInternalDependencies()", () => {
   });
 });
 
-describe("addToQueue", () => {
-  it("adds internal dependencies to the queue", async () => {
-    const packageRoot = await setupFixture("monorepo");
-
-    const packageToAdd = "package-a";
-    const packagePath = path.join(packageRoot, "packages", packageToAdd);
-
-    const queue: string[] = [];
-    const done: PackageHashInfo[] = [];
-
-    addToQueue(packageToAdd, packagePath, queue, done);
-
-    const expectedQueue = [packagePath];
-    expect(queue).toEqual(expectedQueue);
-  });
-
-  it("doesn't add to the queue if the package has been evaluated", async () => {
-    const packageRoot = await setupFixture("monorepo");
-
-    const packageToAdd = "package-a";
-    const packagePath = path.join(packageRoot, "packages", packageToAdd);
-
-    const queue: string[] = [];
-    const done: PackageHashInfo[] = [
-      { name: packageToAdd, filesHash: "", dependenciesHash: "" }
-    ];
-
-    addToQueue(packageToAdd, packagePath, queue, done);
-
-    expect(queue).toEqual([]);
-  });
-
-  it("doesn't add to the queue if the package is already in the queue", async () => {
-    const packageRoot = await setupFixture("monorepo");
-
-    const packageToAdd = "package-a";
-    const packagePath = path.join(packageRoot, "packages", packageToAdd);
-
-    const queue: string[] = [packagePath];
-    const done: PackageHashInfo[] = [];
-
-    addToQueue(packageToAdd, packagePath, queue, done);
-
-    const expectedQueue = [packagePath];
-    expect(queue).toEqual(expectedQueue);
-  });
-});
-
 describe("resolveInternalDependencies()", () => {
   it("adds internal dependency names to the processedPackages list", async () => {
     const packageRoot = await setupFixture("monorepo");
@@ -85,14 +34,9 @@ describe("resolveInternalDependencies()", () => {
 
     const dependencies = { "package-a": "1.0.0", foo: "1.0.0" };
 
-    const queue: string[] = [];
-    const done: PackageHashInfo[] = [];
-
     const resolvedDependencies = resolveInternalDependencies(
       dependencies,
-      workspaces,
-      queue,
-      done
+      workspaces
     );
 
     expect(resolvedDependencies).toEqual(["package-a"]);
