+++ packages/hasher/src/resolveInternalDependencies.ts
@@ -1,9 +1,4 @@
-import { PackageHashInfo } from "./hashOfPackage";
-import {
-  WorkspaceInfo,
-  listOfWorkspacePackageNames,
-  findWorkspacePath
-} from "./yarnWorkspaces";
+import { WorkspaceInfo, listOfWorkspacePackageNames } from "./yarnWorkspaces";
 
 export type Dependencies = { [key in string]: string };
 
@@ -17,43 +12,14 @@ export function filterInternalDependencies(
   );
 }
 
-function isDone(done: PackageHashInfo[], packageName: string): boolean {
-  return Boolean(done.find(({ name }) => name === packageName));
-}
-
-function isInQueue(queue: string[], packagePath: string): boolean {
-  return queue.indexOf(packagePath) >= 0;
-}
-
-export function addToQueue(
-  name: string,
-  dependencyPath: string,
-  queue: string[],
-  done: PackageHashInfo[]
-): void {
-  if (!isDone(done, name) && !isInQueue(queue, dependencyPath)) {
-    queue.push(dependencyPath);
-  }
-}
-
 export function resolveInternalDependencies(
   allDependencies: Dependencies,
-  workspaces: WorkspaceInfo,
-  queue: string[],
-  done: PackageHashInfo[]
+  workspaces: WorkspaceInfo
 ): string[] {
   const dependencyNames = filterInternalDependencies(
     allDependencies,
     workspaces
   );
-
-  dependencyNames.forEach(name => {
-    const dependencyPath = findWorkspacePath(workspaces, name);
-
-    if (dependencyPath) {
-      addToQueue(name, dependencyPath, queue, done);
-    }
-  });
 
   return dependencyNames;
 }
